requirements.txt
Pillow>=9.1.1
tensorflow>=2.10.0
tqdm>=4.64.1
albumentations>=1.3.0